First bug fix in encode(areaOfInterest): return also the cells that are on the left side of UTM zones.
git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK8@1786516 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-referencing/src/main/java/org/apache/sis/internal/referencing/provider/PolarStereographicA.java
@@ -111,8 +111,10 @@ public final class PolarStereographicA extends AbstractStereographic {
 
     /**
      * False Easting and false Northing value used in Universal Polar Stereographic (UPS) projections.
+     * Represented as an integer for the convenience of Military Reference Grid System (MGRS) or other
+     * grid systems.
      */
-    public static final double UPS_SHIFT = 2000000;
+    public static final int UPS_SHIFT = 2000000;
 
     /**
      * Sets the parameter values for a Universal Polar Stereographic projection
